refactor: clear compiler deprecation and lint warnings
- hiltViewModel: import from androidx.hilt.lifecycle.viewmodel.compose
- MenuAnchorType -> ExposedDropdownMenuAnchorType
- TabRow -> PrimaryTabRow (add-torrent and torrent-details tabs)
- Icons.Filled.Sort -> AutoMirrored; plain-tooltip position provider ->
  rememberTooltipPositionProvider(TooltipAnchorPosition.Above)
- drop an unnecessary !! in BackupManager and a redundant toFloat() in TorrentRow

+++ app/src/main/kotlin/dev/yashgarg/qbit/ui/server/TorrentRow.kt
@@ -143,7 +143,7 @@ fun TorrentRow(
                 }
 
                 LinearProgressIndicator(
-                    progress = { torrent.progress.toFloat() },
+                    progress = { torrent.progress },
                     modifier = Modifier.fillMaxWidth().padding(vertical = 6.dp),
                     color = visuals.progressColor,
                     trackColor = MaterialTheme.colorScheme.surfaceVariant,
